requirements.txt

streamlit
tensorflow
keras
gdown
numpy
Pillow
